src/components/Button/button.test.tsx
import React from "react";
import { render } from '@testing-library/react'

import Button from "./button";

test('our first test for Button Component', () => {
    const wrapper = render(<Button>Nice</Button>)
    const element = wrapper.queryByText('Nice')
    
    expect(element).toBeTruthy()
    expect(element).toBeInTheDocument()
})

describe('test Button component', () => {
    it('should render the correct default button', () => {
        const wrapper = render(<Button>Nice</Button>)
        const element = wrapper.getByText('Nice')
        
        expect(element).toBeInTheDocument()
        expect(element.tagName).toEqual('BUTTON')
        expect(element).toHaveClass('btn btn-default')
    })


})